Backward slicer supports operation with AND

+++ src/Core/Lib/Bits.cs
@@ -35,7 +35,7 @@ namespace Reko.Core.Lib
             return w != 0 && (w & (w - 1)) == 0;
         }
 
-        public static bool IsEvenPowerOfTwo(int n)
+        public static bool IsEvenPowerOfTwo(long n)
         {
             return n != 0 && (n & (n - 1)) == 0;
         }
